+++ src/Mages.Core/Tokens/StringTokenizer.cs
@@ -81,7 +81,7 @@
 
             public IToken Error()
             {
-                AddError(new ParseError(ErrorCode.StringMismatch, _scanner.Position.ToRange()));
+                AddError(ErrorCode.StringMismatch, _scanner.Position.ToRange());
                 return Emit();
             }
 
@@ -105,7 +105,7 @@
                         case CharacterTable.QuestionMark: _buffer.Append('?'); _scanner.MoveNext(); break;
                         case CharacterTable.SingleQuotationMark: _buffer.Append('\''); _scanner.MoveNext(); break;
                         case CharacterTable.DoubleQuotationMark: _buffer.Append('\"'); _scanner.MoveNext(); break;
-                        default: AddError(new ParseError(ErrorCode.EscapeSequenceInvalid, _scanner.Position.ToRange())); break;
+                        default: AddError(ErrorCode.EscapeSequenceInvalid, _scanner.Position.ToRange()); break;
                     }
 
                     if (_scanner.Current != CharacterTable.End)
@@ -133,7 +133,7 @@
                     }
                 }
 
-                AddError(new ParseError(ErrorCode.AsciiSequenceInvalid, _scanner.Position.From(start)));
+                AddError(ErrorCode.AsciiSequenceInvalid, _scanner.Position.From(start));
                 return String.Empty;
             }
 
@@ -146,7 +146,7 @@
                 {
                     if (!_scanner.MoveNext() || !_scanner.Current.IsHex())
                     {
-                        AddError(new ParseError(ErrorCode.UnicodeSequenceInvalid, _scanner.Position.From(start)));
+                        AddError(ErrorCode.UnicodeSequenceInvalid, _scanner.Position.From(start));
                         return String.Empty;
                     }
 
@@ -164,14 +164,14 @@
                 return new StringToken(content, _errors, _start, _scanner.Position);
             }
 
-            private void AddError(ParseError error)
+            private void AddError(ErrorCode code, ITextRange range)
             {
                 if (_errors == null)
                 {
                     _errors = new List<ParseError>();
                 }
 
-                _errors.Add(error);
+                _errors.Add(new ParseError(code, range));
             }
         }